backend_api/requirements.txt
Flask==3.0.2
mysql-connector-python==9.0.0
gunicorn==21.2.0
